Trigger variation gallery migration on feature enable
Move the legacy AVI gallery migration off WC's version-based update
queue and onto the variation gallery feature flag. Running it on
version bump creates a snapshot of state at update time even for
merchants who never enable the feature, and re-running for opt-ins
later isn't tied to anything meaningful.

The migration function moves out of wc-update-functions.php into a
new Migration class in src/Internal/VariationGallery/, where it can
live next to LegacyVariationGalleryCompatibility and Telemetry. The
function itself was already idempotent and AVI-aware; only the
trigger changes.

Package::init() now schedules Migration::run via WC's existing
woocommerce_run_update_callback plumbing when the feature is on,
guarded by a completion option and a duplicate-action search.
Scheduling is deferred to WP's init hook so Action Scheduler's data
store is ready (calling WC()->queue() on plugins_loaded silently
fails with an "as_get_scheduled_actions was called before the
data store was initialized" notice).

Tests follow the source: the migration tests move to
tests/php/src/Internal/VariationGallery/MigrationTest.php, mirroring
the new class location.
